requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
asyncpg==0.30.0
certifi==2024.8.30
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
fastapi==0.115.0
filelock==3.16.1
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
identify==2.6.1
idna==3.10
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
nodeenv==1.9.1
platformdirs==4.3.6
pre-commit==3.8.0
pycodestyle==2.12.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
setuptools==75.1.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typing_extensions==4.12.2
uvicorn==0.32.1
virtualenv==20.26.6
